debugging mechanism
set the attribute `debug` on a <template> to turn on this feature.

when debugging is on, elements with a `print` attribute set will log the
key and value associated with that element.

the `debug` attribute can have a value, a string selector (e.g. "#log")
referring to the log target.

for example, consider `{ "greeting": "hello" }` (hello.json) applied to
the following template.

    <template debug data-src="hello.json">
      <p slot="greeting" print></p>
    </template>

executing the render will emit

    console.log("greeting", { "greeting": "hello" })